+++ CHANGELOG.md
@@ -10,8 +10,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Version `0.5.2` included an error with the `mdx` files for embedding generation from `generateFromFile` API. This has been fixed.
+- Fixed several typos related to the `aggregateMaxTokenNumber` configurable parameter.s
+
+### Changed
+
+- Updated documentation related to the Aggregate Max Token Number and custom prompts (both system and user prompts)
 
 ## 0.5.2 - 2025-01-29
+
 
 ### Fixed